% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siftr-helper.r
\name{addLogs}
\alias{addLogs}
\title{Helper function for pseudo_diri, adds two logs}
\usage{
addLogs(lx, ly)
}
\arguments{
\item{lx}{first value}

\item{ly}{second value}
}
\description{
Helper function for pseudo_diri, adds two logs
}

