% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siftr-helper.r
\name{checkUnequalSequenceLengths}
\alias{checkUnequalSequenceLengths}
\title{Check lengths of sequences are equal}
\usage{
checkUnequalSequenceLengths(aln)
}
\arguments{
\item{aln}{alignment list of sequences}
}
\value{
will throw error if there is at least one unequal sequence length
otherwise, will return length of the query
}
\description{
This checks that the lengths of each sequence in the
alignment is equal to the length of the query
}

