% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siftr-main.r
\name{diffAA.siftr_matrix}
\alias{diffAA.siftr_matrix}
\title{Number of unique amino acids per position}
\usage{
\method{diffAA}{siftr_matrix}(sift_mat)
}
\arguments{
\item{sift_mat}{siftr matrix generated using predictFromAlignment}
}
\value{
integer vector, each value is between 0-20
}
\description{
Number of unique amino acids per position
}

