% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_proc_sift.R
\name{filter_proc_sift}
\alias{filter_proc_sift}
\title{Filter Processed SIFT-MS Data}
\usage{
filter_proc_sift(df, table_type, remove_incomplete_scans = TRUE, conc_unit)
}
\arguments{
\item{df}{A data frame outputted from the read_proc_sift function containing
Processed SIFT-MS data, which includes columns \code{table}, \code{time_s},
\code{compound_identity}, and \code{conc}.}

\item{table_type}{A character string indicating the type of table to filter.
Acceptable values include: \code{"analyte_conc"}, \code{"conc_per_reagent"},
\code{"conc_per_product"}, \code{"raw_analyte_conc"}, \code{"raw_conc_per_reagent"}, and
\code{"raw_conc_per_product"}. Raw tables do not have the ICF correction applied.}

\item{remove_incomplete_scans}{A logical value. If TRUE (default), scans
which are incomplete will be removed. Incomplete scans can result when
a scan is ended.}

\item{conc_unit}{A character string specifying the desired concentration unit for
conversion. Acceptable volumetric units are \code{"ppm"}, \code{"ppb"}, and \code{"ppt"}; and
mass based units \code{"mg/m³"}, \code{"µg/m³"}, and \code{"ng/m³"}. The function does not convert
between mass and volumetric - ensure the correct type when exporting from LabSyft.}
}
\value{
A list with two elements:
\describe{
\item{data}{A data frame with processed and converted concentration values in the requested unit.}
\item{removed_conflicts}{A data frame of rows removed due to volumetric/mass unit conflicts.}
}
}
\description{
This function is designed to filter processed SIFT-MS data outputted from the
read_proc_sift function. It allows you to filter the data frame for one of six
tables, can filter to remove incomplete scans and sets the units for
concentration.
}
