% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siftr-helper.r
\name{getEpsilonValues}
\alias{getEpsilonValues}
\title{Compute epsilon values}
\usage{
getEpsilonValues(pb_weights, pb_weights_norm, rank_matrix)
}
\arguments{
\item{pb_weights}{see pbWeights}

\item{pb_weights_norm}{see pbWeightsNorm}

\item{rank_matrix}{precomputed rank matrix for amino acids (BLOSUM62)}
}
\description{
Used for computing final sift scores
}

