% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siftr-main.r
\name{infoContent.siftr_matrix}
\alias{infoContent.siftr_matrix}
\title{Position information content}
\usage{
\method{infoContent}{siftr_matrix}(sift_mat)
}
\arguments{
\item{sift_mat}{siftr matrix generated using predictFromAlignment}
}
\value{
numeric vector of information contents
}
\description{
Position information content
}

