% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_lastsift.R
\name{lastsift}
\alias{lastsift}
\title{View all matches from the most recent sift}
\usage{
lastsift()
}
\value{
Invisibly returns a dataframe.
}
\description{
This stores the last sift result so that it's accessible even if you've done
other things since then.
}
\details{
It returns the complete table of matches, i.e. even if there were too many
results to print to the console, all results are visible here. If run in
RStudio, it will open the results in the Viewer pane so that you can search,
filter, resort, etc.
}
