% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siftr-helper.r
\name{letterMatrix}
\alias{letterMatrix}
\title{Compute matrix of letters for the alignment}
\usage{
letterMatrix(aln, alphabet, npos)
}
\arguments{
\item{aln}{alignment list of sequences}

\item{alphabet}{character vector of alphabet (amino acids used)}

\item{npos}{number of positions in the alignment}
}
\value{
matrix of amino acid indices for each sequence/position
}
\description{
This takes all n sequences, of m positions and constructs
a matrix containing sequence amino acid x position
amino acids are represented as integers 1-20, 21 non-AA character
}

