% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siftr-helper.r
\name{makePWM}
\alias{makePWM}
\title{Make PWM using pb weights and amino acid frequencies}
\usage{
makePWM(aln_mat, pb_weights, frq)
}
\arguments{
\item{aln_mat}{alignment matrix generated from letterMatrix}

\item{pb_weights}{weights computed from pbWeights function}

\item{frq}{precomputed amino acid frequency in the proteome}
}
\value{
weight matrix of 20 amino acids x number of positions
}
\description{
Computes a weight matrix of 20 amino acids x n positions
Each cell is the frequency of the corresponding amino acid in that position * weight
}

