% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siftr-helper.r
\name{makeSiftMatrix}
\alias{makeSiftMatrix}
\title{Computes final matrix of sift scores returned to the user}
\usage{
makeSiftMatrix(pb_weights, pb_weights_norm, eps, diri)
}
\arguments{
\item{pb_weights}{see pbWeights}

\item{pb_weights_norm}{see pbWeightsNorm}

\item{eps}{epsilon values}

\item{diri}{precomputed dirichelet components, used to compute pseudocounts}
}
\value{
numeric matrix with amino acid rows x number of positions containing sift scores
}
\description{
Computes final matrix of sift scores returned to the user
}

