% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siftr-helper.r
\name{medianInfoContent}
\alias{medianInfoContent}
\title{Get median information content for each position}
\usage{
medianInfoContent(aln_mat, frq, cores = 1)
}
\arguments{
\item{aln_mat}{alignment matrix generated from letterMatrix}

\item{frq}{precomputed amino acid frequency in the proteome}

\item{cores}{number of cores for parallel processing (default: 1)}
}
\value{
Numeric vector with median information content for each position
}
\description{
Get median information content for each position
}

