% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siftr-helper.r
\name{medianOfArray}
\alias{medianOfArray}
\title{SIFT version of median}
\usage{
medianOfArray(arr)
}
\arguments{
\item{arr}{vector of numeric values}
}
\value{
median value
}
\description{
The way medians are computed in sift C code is a bit different
This implements the median function used by sift
}

