% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{mtcars_lab}
\alias{mtcars_lab}
\title{Labelled version of mtcars}
\format{
A dataframe with 13 columns and 32 rows.
}
\source{
\code{mtcars}
}
\usage{
mtcars_lab
}
\description{
This is mtcars with value labels, variable labels (in \code{vs} only), some transformation
to factor (\code{car} and \code{am}), and an added Logical column (\code{above_avg}).
}
\keyword{datasets}
