% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siftr-helper.r
\name{normalizeWeights}
\alias{normalizeWeights}
\title{Normalize pb weights matrix}
\usage{
normalizeWeights(weights)
}
\arguments{
\item{weights}{weights computed from pbWeights function}
}
\value{
normalized pb weight matrix
}
\description{
This takes wight matrix produced by pbWeights and normalizes
by number of sequences in alignment / sum of all weights
}

