% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siftr-helper.r
\name{pbWeights}
\alias{pbWeights}
\title{Compute a weight for each amino acid at each position}
\usage{
pbWeights(aln_mat, return_all = FALSE)
}
\arguments{
\item{aln_mat}{alignment matrix generated from letterMatrix}

\item{return_all}{if TRUE, returns intermediate variables computed (pfm, diff_aas, etc), otherwise just weight matrix}
}
\value{
numeric matrix of n sequences x m positions, with a weight for each cell
}
\description{
These weights are used to compute a sequence weight
denoting how much information each sequence is contributing
}

