% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siftr-helper.r
\name{pbWeightsNorm}
\alias{pbWeightsNorm}
\title{Normalizes weights computed by pbWeights}
\usage{
pbWeightsNorm(pb_weights, aln_mat)
}
\arguments{
\item{pb_weights}{weights computed from pbWeights function}

\item{aln_mat}{alignment matrix generated from letterMatrix}
}
\value{
list of two elements: (1) matrix of normalized weights
grouped by amino acids (20 x number of positions)
(2) numeric vector of the sum of weights per positions
}
\description{
This normalizes the pbWeights weight matrix using normalizeWeights
then groups by amino acid for 20 x number of position matrix
This is used to compute sume of weights for each position
}

