% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siftr-helper.r
\name{percentIdentity}
\alias{percentIdentity}
\title{Compute percent identity between two sequences}
\usage{
percentIdentity(comp, seq)
}
\arguments{
\item{comp}{sequence being compared}

\item{seq}{suery sequence}
}
\value{
percent identity between both sequences
}
\description{
Compute percent identity between two sequences
}

