% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siftr-helper.r
\name{positionInfoContent}
\alias{positionInfoContent}
\title{Get median information content for a given position}
\usage{
positionInfoContent(position = 1, aln_mat, frq, small_sample_correction = F)
}
\arguments{
\item{position}{alignment position of interest}

\item{aln_mat}{alignment matrix generated from letterMatrix}

\item{frq}{precomputed amino acid frequency in the proteome}

\item{small_sample_correction}{TRUE if small sample correction is applied
to information content calculation (not implemented currently, sift has it but doesn't use it)}
}
\value{
information content value between 0 (conserved) to log2(20) (least conserved)
}
\description{
Get sequences that have no gap or X in position i,
compute median information content accross all positions
}

