% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siftr-main.r
\name{predictFromAlignment}
\alias{predictFromAlignment}
\title{Predict sift scores from a protein alignment}
\usage{
predictFromAlignment(aln, cores = 1)
}
\arguments{
\item{aln}{path to alignment file in fasta format or character vector of protein sequences}

\item{cores}{number of cores for parallel processing (default: 1)}
}
\value{
matrix of sift scores
}
\description{
Predict sift scores from a protein alignment
}
\examples{
# Generate dummy alignment
aln = c('SSSS', 'STTT', 'SYSY', 'ASST', 'KKHS')

# Generate matrix of sift scores
sift_mat = predictFromAlignment(aln)

# Show summary of run
summary(sift_mat)

# Convert to data frame and filter out unreliable predictions
sift_df = filterPredictions(sift_mat, score_thresh = 0.05, residue_thresh = 2)
}

