% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siftr-main.r
\name{print.siftr_matrix}
\alias{print.siftr_matrix}
\title{Print siftr matrix}
\usage{
\method{print}{siftr_matrix}(x, list_attributes = FALSE, ...)
}
\arguments{
\item{x}{object of class siftr_matrix}

\item{list_attributes}{if TRUE will show stored attributes for object}
}
\description{
Print siftr matrix
}

