% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fullscan_sift.R
\name{read_fullscan_sift}
\alias{read_fullscan_sift}
\title{Read SIFT Full Scan Data}
\usage{
read_fullscan_sift(file, drop_prep = F, chatty = T, warn = FALSE)
}
\arguments{
\item{file}{The file path for the SIFT .csv file.}

\item{drop_prep}{Is the PREPARATION sub-data missing? TRUE/FALSE.}

\item{chatty}{\code{TRUE}/\code{FALSE}. Should the function communicate what it is doing? Useful for debugging.}

\item{warn}{logical. if a section of the file cannot be read, should a warning or error be thrown? Default FALSE leads to an error being produced. For use with \code{read_many_sift()} to skip bad files}
}
\value{
A tibble containing processed time vs mass data.
}
\description{
This function reads extracts the "Time vs Mass" table from a SIFT-MS file. Useful for fullscan SIFT-MS data which is missing several tables usually found in a SIM SIFT-MS file and causes an error in read_sift.
}
