% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fullscan_xml_sift.R
\name{read_fullscan_xml_sift}
\alias{read_fullscan_xml_sift}
\title{Read Full-Scan XML SIFT Data}
\usage{
read_fullscan_xml_sift(xml_file_path, tzone = "Europe/London")
}
\arguments{
\item{xml_file_path}{Character. Path to the XML file to be read.}

\item{tzone}{Character. Time zone for parsing timestamps. Defaults to \code{Europe/London}.}
}
\value{
A named list containing the following elements:
\item{settings}{Data frame of settings extracted from the XML file.}
\item{method}{Data frame of method values extracted from the XML file.}
\item{scan_masses}{Data frame of masses scanned}
\item{measurements}{Data frame containing fullscan data}
}
\description{
Parses a full-scan SIFT XML file and extracts information including settings, methods,
scan masses, and measurements. The data is returned as a list of data frames.
}
