% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_many_fullscan_xml_sift.R
\name{read_many_fullscan_xml_sift}
\alias{read_many_fullscan_xml_sift}
\title{Read and Combine Multiple Fullscan XML SIFT Files}
\usage{
read_many_fullscan_xml_sift(xml_file_paths, tzone = "Europe/London")
}
\arguments{
\item{xml_file_paths}{A character vector of XML file paths to be processed.}

\item{tzone}{A character string specifying the time zone for parsing the file timestamps.
Default is "Europe/London".}
}
\value{
A list containing:
\describe{
\item{settings}{Combined settings data from all successfully processed XML files.}
\item{method}{Combined method data from all successfully processed XML files.}
\item{scan_masses}{Combined scan mass data from all successfully processed XML files.}
\item{measurements}{Combined measurements data from all successfully processed XML files.}
\item{errors}{A dataframe listing files that failed to process and their error messages.}
}
}
\description{
This function reads multiple XML files using \code{read_fullscan_xml_sift} and combines the resulting dataframes
from each file into five tables: \code{settings}, \code{method}, \code{scan_masses}, \code{measurements} and \code{errors}
}
