% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_many_sift.R
\name{read_many_sift}
\alias{read_many_sift}
\title{Read Many SIFT Files}
\usage{
read_many_sift(files, table = "concentrations", ...)
}
\arguments{
\item{files}{A vector of file paths.}

\item{table}{A table from the SIFT data. One of "time", "meta", "prep_phase",
"sample_phase", "phase_mean_values", "intensity_corrected", "time_vs_mass",
"concentrations", "analytes" or "summary". Defaults to "concentrations".}

\item{...}{Arguments to pass to "read_sift"}
}
\value{
A tibble.
}
\description{
This function is a wrapper around "read_sift" to read in
multiple SIFT .csv files. The user must specify a specific table within the
SIFT data (defaulting to "concentrations") and the function will output a
single tibble with that table from all provided .csv files, tagged with the
start time from each.
}
