% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_many_sift2.R
\name{read_many_sift2}
\alias{read_many_sift2}
\title{Read Many SIFT Files}
\usage{
read_many_sift2(files, ...)
}
\arguments{
\item{files}{A vector of file paths.}

\item{...}{Arguments to pass to "read_sift"}
}
\value{
A list of tibbles.
}
\description{
This function is an alternative wrapper around "read_sift" to
read in multiple SIFT .csv files. Unlike "read_many_sift", this function
does not require specifying a table as it reads in all of them. All files
are bound and labelled with the starting time of the run.
}
