% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siftr-helper.r
\name{removeIdenticalSeqs}
\alias{removeIdenticalSeqs}
\title{Remove identical sequences in a list alignment}
\usage{
removeIdenticalSeqs(aln)
}
\arguments{
\item{aln}{alignment list of sequences}
}
\value{
alignment list with identical sequences removed
}
\description{
Given an alignment of sequences (as a list),
Remove sequences that have 100% identity to query (first element)
}

