% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sent_sif.R
\name{sent_sif}
\alias{sent_sif}
\title{Calculates SIF weighted embedding for given sentence. Uses word embeddings calculated via word_sif function}
\usage{
sent_sif(sentence = "I need a chicken tender")
}
\arguments{
\item{sentence}{character: target sentence}
}
\value{
numeric vector: weighted and averaged embedding vector for target sentence
}
\description{
Calculates SIF weighted embedding for given sentence. Uses word embeddings calculated via word_sif function
}
