% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sift-package.R
\docType{package}
\name{sift-package}
\alias{{pkgname}-package}
\title{sift: Fuzzily search dataframe columns, labels, and factor levels to find variables}
\description{
Fuzzily search dataframe column names, labels, and factor levels to find which columns contain the variables you need. Designed for use with large dataframes that can span thousands of columns and millions of rows.
}
\keyword{internal}
