% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{stopwords_sift}
\alias{stopwords_sift}
\title{List of stopwords for making terse labels}
\format{
A character vector with 338 elements.
}
\source{
\url{https://github.com/koheiw/stopwords}
}
\usage{
stopwords_sift
}
\description{
This is a list of stopwords that was generated from \code{stopwords::stopwords(source = src)}
where \code{src} is one of \code{c("snowball", "stopwords-iso", "smart", "marimo", "nltk")} and
language is \code{"en"}. I kept any word that appeared in at least 3 out of 5 of those lists,
and then re-added these words that are pertinent to data analysis: "until", "while",
"against", "between", "during", "before", "after", "above", "below", "up", "down", "no",
"on", "off", "over", "under", "once", "when", "where", "why", "how", "all", "any",
"both", "each", "few", "more", "most", "not", "only", "same", "again", "many", "who". I
duplicated the list and made the first half in lowercase, and the second half in sentence
case.
}
\keyword{datasets}
