% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siftr-main.r
\name{summary.siftr_matrix}
\alias{summary.siftr_matrix}
\title{Summary for siftr matrix}
\usage{
\method{summary}{siftr_matrix}(sift_mat)
}
\arguments{
\item{sift_mat}{siftr matrix generated using predictFromAlignment}
}
\description{
Summary for siftr matrix
}

