% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_as_quantitative_sift.R
\name{validate_as_quantitative_sift}
\alias{validate_as_quantitative_sift}
\title{Validate SIFT data as quantitative}
\usage{
validate_as_quantitative_sift(data)
}
\arguments{
\item{data}{A data frame containing fullscan data SIFT including columns
\code{ion}, \code{date}, and \code{value}.}
}
\value{
A list of tibbles. \code{validated_data} is a data frame with additional
columns (\code{product_ion_mass}, \code{reagent_ion_mass}, \verb{total_ion _count},
\code{reagent_ion_count}, \code{sum_product_ions}, \code{ratio_reagent_to_product} and
\code{quantitative}) containing validation information, and there are two other
data frames to aid with potential warning messages.
}
\description{
This function validates that fullscan SIFT-MS data is quantitative through
addition of a logical column. It ensures that the sum of product ions for a
particular reagent ion and date (dttm) do not exceed that of the reagent ion
(for the H30+ ion this includes the first water cluster). This function is
designed for data obtained with the H3O+, NO+ and/or O2+ reagent ions.
}
