% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_sif.R
\name{word_sif}
\alias{word_sif}
\title{Calculates SIF weighted embedding for a single word. Not particularly useful on its own, as SIF assumes multiple word embeddings to average together}
\usage{
word_sif(word = "chicken")
word_sif(word = "tender", efl = ef_list, weight_param = .004)
}
\arguments{
\item{word}{character: target word}

\item{efl}{list: embedding and frequency object, assumes each word-named list entry has freq and embedding}

\item{weight_param}{numeric: parameter used to tweak SIF calculation, default = .001}
}
\value{
numeric vector: weighted embedding vector for target word
}
\description{
Calculates SIF weighted embedding for a single word. Not particularly useful on its own, as SIF assumes multiple word embeddings to average together
}
