% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_date_sift.R
\name{add_date_sift}
\alias{add_date_sift}
\title{Add Date SIFT Function}
\usage{
add_date_sift(data, units, filter_phase = FALSE, force_tz = "Europe/London")
}
\arguments{
\item{data}{A data frame containing the SIFT data to be processed.}

\item{units}{A character string specifying the units for time in the unstitched files. Can be "s" for seconds or "ms" for milliseconds. This argument is required.}

\item{filter_phase}{A character string specifying the phase to filter by. Options are "SAMPLE", "BACKGROUND", "CALIBRATION", or \code{FALSE} to not filter. Defaults to 'FALSE'. Check the 'Phase' column in SIFT files to find correct option; for most users this will be "SAMPLE" for raw SIM or fullscan data and 'FALSE' for processed data.}

\item{force_tz}{A character string containing the time zone to convert to, or \code{FALSE} to not adjust the time zone. R must recognize the name contained in the string as a time zone on your system. Defaults to "Europe/London".}
}
\value{
A data frame with an added \code{date} column, time zone set for date and start time columns, and optionally filtered by phase.
}
\description{
This function adds a date column to SIFT data, sets the timezone of date and start_time columns, and optionally filters by phase.
}
