% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_proc_sift.R
\name{filter_proc_sift}
\alias{filter_proc_sift}
\title{Filter Processed SIFT-MS Data}
\usage{
filter_proc_sift(df, table_type, remove_incomplete_scans = TRUE, conc_unit)
}
\arguments{
\item{df}{A data frame outputted from the read_proc_sift function containing
Processed SIFT-MS data, which includes columns \code{table}, \code{time_s},
\code{compound_identity}, and \code{conc}.}

\item{table_type}{A character string indicating the type of concentration data
to filter. Acceptable values include: "analyte_conc", "conc_per_reagent",
"conc_per_product", "raw_analyte_conc", "raw_conc_per_reagent", and
"raw_conc_per_product"}

\item{remove_incomplete_scans}{A logical value. If TRUE (default), scans
which are incomplete will be removed. Incomplete scans can result when
a scan is ended.}

\item{conc_unit}{A character string specifying the desired concentration unit for
conversion. Acceptable values are "ppm", "ppb", or "ppt".}
}
\value{
A data frame
}
\description{
This function is designed to filter processed SIFT-MS data outputted from the
read_proc_sift function. It allows you to filter the data frame for one of six
tables, can filter to remove incomplete scans and sets the units for
concentration.
}
