% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_sift.R
\name{options_sift}
\alias{options_sift}
\title{Set and get options related to how sift() runs.}
\usage{
options_sift(
  key = c("sift_loudly", "sift_terse", "sift_limit", "sift_guessmax"),
  val = NULL
)
}
\arguments{
\item{key}{(String) The name of an option.}

\item{val}{(Optional) A new value for the option, if you want to change it.}
}
\value{
The option's value.
}
\description{
\itemize{
\item \code{sift_loudly} (Logical; default \code{TRUE}) --- Should optional informative messages be shown?
\item \code{sift_stopwords} (Logical; default \code{TRUE}) --- Should stopwords be removed from \code{sift()}
console output? This makes the output terser, so it's good for removing low-value
words from wordy labels. The list of stopwords is an included dataset called
\code{stopword_sift}.
\item \code{sift_limit} (Integer; default \code{10}) --- How many matches should \code{sift()} print? This
is meant to save you from accidentally printing a summary that is hundreds of columns long.
\item \code{sift_guessmax} (Integer; default \code{1000}) --- What is the maximum number of elements that
gets randomly sampled from each column to build summary stats? This includes stats like
frequent values and random previews.
}
}
\examples{
# options_sift("sift_loudly")  # Returns the option's current value

# options_sift("sift_loudly", FALSE)  # Change the value to something else.

# Options set in this function are set in R's options() interface
# options("sift_loudly")
# getOption("sift_loudly")
}
