% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_proc_sift.R
\name{read_proc_sift}
\alias{read_proc_sift}
\title{Read Processed SIFT Data}
\usage{
read_proc_sift(file)
}
\arguments{
\item{file}{The file path for the pre-processed SIFT .csv file.}
}
\value{
A list.
}
\description{
This is a function to read data that has already been processed
by the SYFT into three tables - analyte concentrations, concentrations per
reagent and concentrations per product.
}
