% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sift.R
\name{read_sift}
\alias{read_sift}
\title{Read SIFT Data}
\usage{
read_sift(file)
}
\arguments{
\item{file}{The file path for the SIFT .csv file.}
}
\value{
A list.
}
\description{
Raw SIFT data is very messy and hard to use. This function reads SIFT data from a .csv and "tidies" each element to enhance usability. The output is a list of each individual table that the SIFT outputs, as well as the date and time the SIFT began working.
}
