% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sift.R
\name{read_sift}
\alias{read_sift}
\title{Read SIFT Data}
\usage{
read_sift(file, drop_prep = F, chatty = T)
}
\arguments{
\item{file}{The file path for the SIFT .csv file.}

\item{drop_prep}{Is the PREPARATION sub-data missing? TRUE/FALSE.}

\item{chatty}{\code{TRUE}/\code{FALSE}. Should the function communicate what
it is doing? Useful for debugging.}
}
\value{
A list.
}
\description{
Raw SIFT data is very messy and hard to use. This function reads
SIFT data from a .csv and "tidies" each element to enhance usability. The
output is a list of each individual table that the SIFT outputs, as well as
the date and time the SIFT began working. Also has functionality to specify
if the PREPARATION PHASE is present or not.
}
